Specialized Literal garbage collector
This was because of the lFunction class holding a reference to a block.
That block would be deleted twice, so now the literal garbage collector
sets the pointer to the block to nullptr. It's assumed that the block
itself will be cleared by a call to a different garbage collector.